Implementación Mantenimiento Pacientes y Medicamentos

+++ SIMCIC/SIMCIC.DATA/Paciente.cs
@@ -8,13 +8,13 @@ namespace SIMCIC.DATA
 {
     public class Paciente
     {
-        public int Cedula { get; set; }
+        public Int64 Cedula { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public string FechaNacimiento { get; set; }
         public int Edad { get; set; }
         public string Sexo { get; set; }
-        public int Telefono { get; set; }
+        public Int64 Telefono { get; set; }
         public string Correo { get; set; }
         public string Ocupacion { get; set; }
     }
